dvfs: Extract super_block to a separate file
