+++ src/fs/dvfs/Mybuild
@@ -29,6 +29,7 @@ module core extends embox.fs.fs_api {
 	source "dvfs_dentry.c"
 	source "dvfs_idesc_ops.c"
 	source "dvfs_util.c"
+	source "super_block.c"
 	source "inode.c"
 
 	depends embox.fs.path_helper
